requirements.txt
google-adk[database,eval]
openinference-instrumentation-google-adk
arize-otel
yfinance
psutil
litellm>=1.75.5
python-dotenv>=1.0.0
Deprecated
pytest>=8.3.4
fastmcp>=2.3.4
uvicorn>=0.34.0
mypy>=1.15.0
ruff
black
httpx
google-genai>=1.0.0
click>=8.1.8
a2a-sdk>=0.3.4
nest-asyncio
requests>=2.32.4
